fix: separated app and index
